Move regular expression parser to new source file
